+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Release History
 
 [Communication UI Calling Library CHANGELOG](docs/CHANGELOG_UI_CALLING.md)
+
+### Features
+- New error code 'unknownError' introduced for ambiguous errors such as device manager error